models/doctr-api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pillow>=10.0.0
pydantic>=2.0.0
numpy>=1.24.0

# DocTR (PyTorch backend)
python-doctr[torch]>=0.9.0
